# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/tasks": {
      "get": {
        "summary": "Get all tasks for the authenticated user",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "status",
                  "priority"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string"
                  },
                  "due_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "assigned_to": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}